added all additional pages, connected the pages together using session to share data.
