+++ assignment1/pom.xml
@@ -20,6 +20,12 @@
   </properties>
 
   <dependencies>
+  <dependency>
+    <groupId>javax.servlet</groupId>
+    <artifactId>javax.servlet-api</artifactId>
+    <version>4.0.1</version>
+    <scope>provided</scope>
+  </dependency>
     <dependency>
       <groupId>junit</groupId>
       <artifactId>junit</artifactId>
@@ -39,10 +45,11 @@
   </dependencies>
 
   <build>
-
+    <sourceDirectory>src/main/java</sourceDirectory>
     <finalName>assignment1</finalName>
     <pluginManagement><!-- lock down plugins versions to avoid using Maven defaults (may be moved to parent pom) -->
       <plugins>
+
         <plugin>
           <artifactId>maven-clean-plugin</artifactId>
           <version>3.1.0</version>
