+++ assignment1/src/main/java/iotbay/model/Device.java
@@ -1,0 +1,5 @@
+package iotbay.model;
+
+public class Device {
+    
+}
