+++ assignment1/src/main/java/iotbay/model/Order.java
@@ -1,0 +1,5 @@
+package iotbay.model;
+
+public class Order {
+    
+}
