+++ assignment1/src/main/java/iotbay/util/DBConnector.java
@@ -1,0 +1,23 @@
+package iotbay.util;
+
+import java.sql.Connection;
+import java.sql.DriverManager;
+import java.sql.SQLException;
+
+public class DBConnector {
+    private static final String URL = "jdbc:mysql://localhost:3306/iotbay_db";
+    private static final String USERNAME = "your_db_username";
+    private static final String PASSWORD = "your_db_password";
+
+    static {
+        try {
+            Class.forName("com.mysql.cj.jdbc.Driver");
+        } catch (ClassNotFoundException e) {
+            e.printStackTrace();  // Consider proper logging in production
+        }
+    }
+
+    public static Connection getConnection() throws SQLException {
+        return DriverManager.getConnection(URL, USERNAME, PASSWORD);
+    }
+}
