--- assignment1/src/main/java/model/Device.java
@@ -1,5 +1,0 @@
-package model;
-
-public class Device {
-    
-}
