--- assignment1/src/main/java/model/Order.java
@@ -1,5 +1,0 @@
-package model;
-
-public class Order {
-    
-}
