--- assignment1/src/main/java/model/User.java
@@ -1,18 +1,0 @@
-package model;
-
-public class User {
-    private String email;
-    private String fullName;
-    private Integer phoneNumber;
-    private String password;
-    private String address1;
-    private String address2;
-    private String suburb;
-    private Integer postcode;
-    private String state; 
-    private Integer cardNumber;
-    private Integer cardExpiry;
-    private Integer cardCVV;
-
-
-}
