+++ assignment1/src/main/webapp/index.jsp
@@ -1,5 +1,28 @@
-<html>
+<%@ page contentType="text/html;charset=UTF-8" language="java" %>
+<!DOCTYPE html>
+<html lang="en" id="index-page">
+<head>
+    <meta charset="UTF-8">
+    <meta name="viewport" content="width=device-width, initial-scale=1.0">
+    <title>IoTBay Home</title>
+    <link rel="stylesheet" href="style.css">
+</head>
 <body>
-<h2>Hello World!</h2>
+ 
+    <nav class="page-nav">
+       <div class="nav-left">
+          <a href="index.jsp">Home</a>
+       </div>
+       <div class="nav-right">
+          <a href="login.jsp">Login</a>
+          <a href="register.jsp">Register</a>
+       </div>
+    </nav>
+    
+   
+    <div class="index-content">
+        <h2>Hello World!</h2>
+        <p>Welcome to IoTBay</p>
+    </div>
 </body>
 </html>
